move demangler datatypes
